go vet crashes when invoked on cgo code using spacemonkey openssl
When running go vet on a code base using cgo, we get infinite recursion during the check:
```
go vet ./...
```

The same type parameter goes in (0x3f0b20), and infinite recursion ensues:

```
main.typeOKForCgoCall(0x3f0b20, 0xc420493f30, 0xc42046d001)
	/usr/local/go/src/cmd/vet/cgo.go:119 +0x160 fp=0xc440701a38 sp=0xc4407019f8
main.typeOKForCgoCall(0x3f0ae0, 0xc42049cc60, 0xc420493f30)
	/usr/local/go/src/cmd/vet/cgo.go:124 +0x104 fp=0xc440701a78 sp=0xc440701a38
main.typeOKForCgoCall(0x3f0b20, 0xc420493f30, 0xc42046d001)
	/usr/local/go/src/cmd/vet/cgo.go:119 +0x160 fp=0xc440701ab8 sp=0xc440701a78
```

These pointers in this case happen to be a `openssl.Certificate` and a `*openssl.Certificate`,
with Underlying() of `*type.Pointer` and `*type.Struct`.

This function either returns immediately when the answer is yes,
or immediately return no, or recurse when the answer is possibly no.
So if we see the same type twice in the list of parents from which
the recursion began, then the answer must be no - because it will never
change to a yes.

Fixes #18389